+++ src/duk_heap_memory.c
@@ -87,7 +87,7 @@ void *duk_heap_mem_alloc(duk_heap *heap, duk_size_t size) {
 	 */
 
 	if (DUK_HEAP_HAS_MARKANDSWEEP_RUNNING(heap)) {
-		DUK_D(DUK_DPRINT("duk_heap_mem_alloc() failed, gc in progress (gc skipped), alloc size %d", size));
+		DUK_D(DUK_DPRINT("duk_heap_mem_alloc() failed, gc in progress (gc skipped), alloc size %ld", (long) size));
 		return NULL;
 	}
 
@@ -110,13 +110,13 @@ void *duk_heap_mem_alloc(duk_heap *heap, duk_size_t size) {
 
 		res = heap->alloc_func(heap->alloc_udata, size);
 		if (res) {
-			DUK_D(DUK_DPRINT("duk_heap_mem_alloc() succeeded after gc (pass %d), alloc size %d",
-			                 (int) (i + 1), (int) size));
+			DUK_D(DUK_DPRINT("duk_heap_mem_alloc() succeeded after gc (pass %ld), alloc size %ld",
+			                 (long) (i + 1), (long) size));
 			return res;
 		}
 	}
 
-	DUK_D(DUK_DPRINT("duk_heap_mem_alloc() failed even after gc, alloc size %d", (int) size));
+	DUK_D(DUK_DPRINT("duk_heap_mem_alloc() failed even after gc, alloc size %ld", (long) size));
 	return NULL;
 }
 #else  /* DUK_USE_MARK_AND_SWEEP */
@@ -195,7 +195,7 @@ void *duk_heap_mem_realloc(duk_heap *heap, void *ptr, duk_size_t newsize) {
 	 */
 
 	if (DUK_HEAP_HAS_MARKANDSWEEP_RUNNING(heap)) {
-		DUK_D(DUK_DPRINT("duk_heap_mem_realloc() failed, gc in progress (gc skipped), alloc size %d", (int) newsize));
+		DUK_D(DUK_DPRINT("duk_heap_mem_realloc() failed, gc in progress (gc skipped), alloc size %ld", (long) newsize));
 		return NULL;
 	}
 
@@ -218,13 +218,13 @@ void *duk_heap_mem_realloc(duk_heap *heap, void *ptr, duk_size_t newsize) {
 
 		res = heap->realloc_func(heap->alloc_udata, ptr, newsize);
 		if (res) {
-			DUK_D(DUK_DPRINT("duk_heap_mem_realloc() succeeded after gc (pass %d), alloc size %d",
-			                 (int) (i + 1), (int) newsize));
+			DUK_D(DUK_DPRINT("duk_heap_mem_realloc() succeeded after gc (pass %ld), alloc size %ld",
+			                 (long) (i + 1), (long) newsize));
 			return res;
 		}
 	}
 
-	DUK_D(DUK_DPRINT("duk_heap_mem_realloc() failed even after gc, alloc size %d", (int) newsize));
+	DUK_D(DUK_DPRINT("duk_heap_mem_realloc() failed even after gc, alloc size %ld", (long) newsize));
 	return NULL;
 }
 #else  /* DUK_USE_MARK_AND_SWEEP */
@@ -288,7 +288,7 @@ void *duk_heap_mem_realloc_indirect(duk_heap *heap, duk_mem_getptr cb, void *ud,
 	 */
 
 	if (DUK_HEAP_HAS_MARKANDSWEEP_RUNNING(heap)) {
-		DUK_D(DUK_DPRINT("duk_heap_mem_realloc_indirect() failed, gc in progress (gc skipped), alloc size %d", (int) newsize));
+		DUK_D(DUK_DPRINT("duk_heap_mem_realloc_indirect() failed, gc in progress (gc skipped), alloc size %ld", (long) newsize));
 		return NULL;
 	}
 
@@ -331,13 +331,13 @@ void *duk_heap_mem_realloc_indirect(duk_heap *heap, duk_mem_getptr cb, void *ud,
 
 		res = heap->realloc_func(heap->alloc_udata, cb(ud), newsize);
 		if (res) {
-			DUK_D(DUK_DPRINT("duk_heap_mem_realloc_indirect() succeeded after gc (pass %d), alloc size %d",
-			                 (int) (i + 1), (int) newsize));
+			DUK_D(DUK_DPRINT("duk_heap_mem_realloc_indirect() succeeded after gc (pass %ld), alloc size %ld",
+			                 (long) (i + 1), (long) newsize));
 			return res;
 		}
 	}
 
-	DUK_D(DUK_DPRINT("duk_heap_mem_realloc_indirect() failed even after gc, alloc size %d", (int) newsize));
+	DUK_D(DUK_DPRINT("duk_heap_mem_realloc_indirect() failed even after gc, alloc size %ld", (long) newsize));
 	return NULL;
 }
 #else  /* DUK_USE_MARK_AND_SWEEP */
